+++ src/api/preferences/index.ts
@@ -1,3 +1,8 @@
+export { PREFERENCE_SPLIT_MAX_PAYLOAD_BYTES } from "./chunked";
+export type { PreferenceResult, StorageMode } from "./chunked";
+export * from "./deletePreference";
 export * from "./deletePreferences";
+export * from "./getPreference";
 export * from "./getPreferences";
+export * from "./updatePreference";
 export * from "./updatePreferences";